4.21 blogilista

+++ osa4/blogilista/controllers/blogs.js
@@ -14,8 +14,7 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
   
-  const decodedToken = jwt.verify(request.token,
-    process.env.SECRET)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
   
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
@@ -39,6 +38,21 @@ blogsRouter.post('/', async (request, response) => {
 
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id);
+  if (!blogToDelete) {
+    return response.status(404).json({ error: 'blog not found' });
+  }
+
+  if (blogToDelete.user.toString() !== decodedToken.id.toString()) {
+    return response.status(403).json({ error: 'unauthorized' });
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
